utils: Rename Git command helper module
The Git helper package has named modules for object IO, repository
identity, index plumbing, worktree operations, and ref updates. The command
execution module still uses the generic utils.git basename.

The project lacks a module name that says the remaining helper owns command
execution. Keeping the generic basename makes the helper family harder to
scan after the surrounding modules have specific names.

This commit addresses that naming gap by renaming utils.git to
utils.git_command and updating source imports to use the command-specific
module directly.

Subsequent commits will update tests and architecture guards to follow the
renamed command module.
