+++ src/git_stage_batch/batch/query.py
@@ -13,7 +13,7 @@ from .state_refs import (
 )
 from .validation import validate_batch_name
 from ..utils.file_io import read_text_file_contents
-from ..utils.git import run_git_command
+from ..utils.git_command import run_git_command
 from ..utils.paths import get_batch_metadata_file_path
