+++ src/git_stage_batch/batch/state_refs.py
@@ -12,7 +12,7 @@ from .validation import validate_batch_name
 from ..core.buffer import LineBuffer
 from ..utils.repository_buffers import load_git_object_as_buffer
 from ..utils.file_io import read_text_file_contents
-from ..utils.git import (
+from ..utils.git_command import (
     run_git_command,
 )
 from ..utils.git_refs import (
