+++ src/git_stage_batch/batch/submodule_pointer.py
@@ -6,7 +6,7 @@ import shutil
 
 from ..exceptions import exit_with_error
 from ..i18n import _
-from ..utils.git import run_git_command
+from ..utils.git_command import run_git_command
 from ..utils.git_worktree import (
     git_checkout_detached,
     git_submodule_update_checkout,
