+++ src/git_stage_batch/commands/discard.py
@@ -40,7 +40,7 @@ from ..core.buffer import (
 )
 from ..exceptions import exit_with_error
 from ..i18n import _
-from ..utils.git import run_git_command
+from ..utils.git_command import run_git_command
 from ..utils.git_repository import require_git_repository
 from ..utils.journal import log_journal
 from ..utils.paths import (
