+++ src/git_stage_batch/commands/selection/selected_file_discarding.py
@@ -8,7 +8,7 @@ from ...data.selected_change.paths import get_selected_change_file_path
 from ...data.session import snapshot_file_if_untracked
 from ...data.undo import undo_checkpoint
 from ...i18n import _
-from ...utils.git import run_git_command
+from ...utils.git_command import run_git_command
 from ...utils.git_worktree import git_checkout_paths
 from ...utils.git_repository import get_git_repository_root_path
 from .action_completion import finish_selected_change_action
