+++ src/git_stage_batch/commands/status.py
@@ -13,7 +13,7 @@ from ..exceptions import CommandError
 from ..i18n import _
 from ..output.status import print_status_summary as _print_status_summary
 from ..output.status_prompt import prompt_needs_status_summary, render_prompt_status
-from ..utils.git import run_git_command
+from ..utils.git_command import run_git_command
 from ..utils.git_repository import require_git_repository
