+++ src/git_stage_batch/data/file_modes.py
@@ -6,7 +6,7 @@ import os
 from pathlib import Path
 import stat
 
-from ..utils.git import run_git_command
+from ..utils.git_command import run_git_command
 from ..utils.git_repository import get_git_repository_root_path
