+++ src/git_stage_batch/data/undo_refs.py
@@ -7,7 +7,7 @@ from ..batch.ref_names import (
     BATCH_STATE_REF_PREFIX,
     LEGACY_BATCH_REF_PREFIX,
 )
-from ..utils.git import run_git_command
+from ..utils.git_command import run_git_command
 from ..utils.git_refs import update_git_refs
