+++ src/git_stage_batch/utils/git_object_io.py
@@ -6,7 +6,7 @@ import subprocess
 from collections.abc import Iterable, Iterator
 from dataclasses import dataclass
 
-from .git import run_git_command, stream_git_command
+from .git_command import run_git_command, stream_git_command
 
 
 @dataclass(frozen=True)
