+++ src/git_stage_batch/utils/repository_buffers.py
@@ -7,7 +7,7 @@ import subprocess
 from collections.abc import Iterator
 
 from ..core.buffer import LineBuffer
-from ..utils.git import stream_git_command
+from ..utils.git_command import stream_git_command
 from ..utils.git_repository import get_git_repository_root_path
 from ..utils.git_object_io import list_git_tree_blobs, read_git_blob
